--- .github/copilot-instructions.md
@@ -1,8 +1,0 @@
-日本語で応答すること
-
-# プロジェクト概要
-
-- このプロジェクトは、`aexlo-rs`というRust製のライブラリです。
-- `aexlo-rs`は、unsafeなコードを多く使用し、After EffectsのC++ SDKと直接やり取りするための低レベルなバインディングを行います。
-- `aexlo-rs`は、`After Effects SDK`に基づいたインターフェイスを提供し、After Effectsプラグインと相互に作用することを目的としています。
-- `aexlo-rs`は、外部DLLのAfter Effectsプラグインをロードし、実行することを目的としています。
