+++ aexlo/src/core/instance.rs
@@ -1,101 +1,15 @@
-use crate::error::{AexloError, Result};
-use crate::suites::SuiteContainer;
+use crate::core::error::{AexloError, Result};
+// use crate::suites::SuiteContainer; // Not needed
+
 use after_effects::ParamType;
 use after_effects_sys::{
-	A_long, A_u_long, PF_Boolean, PF_CustomUIInfo, PF_Err, PF_Err_NONE, PF_Field, PF_Field_UPPER,
-	PF_InFlag_NONE, PF_InFlags, PF_ParamDef, PF_ParamDefPtr, PF_ParamIndex, PF_ParamType, PF_Pixel,
-	PF_PlatDataID, PF_ProgPtr,
+	PF_Boolean, PF_Err_NONE, PF_Field, PF_Field_UPPER, PF_InFlag_NONE, PF_InFlags, PF_ParamDef,
+	PF_ParamType, PF_Pixel,
 };
 use colored::Colorize;
 use dlopen2::wrapper::{Container, WrapperApi};
 use std::path::{Path, PathBuf};
 use std::ptr::null_mut;
-
-use crate::diagnostics::DiagnosticBuilder;
-
-unsafe extern "C" fn rusty_add_param(
-	effect_ref: PF_ProgPtr,
-	index: PF_ParamIndex,
-	def: PF_ParamDefPtr,
-) -> PF_Err {
-	#[cfg(feature = "diagnostics")]
-	DiagnosticBuilder::new()
-		.set_name("PF Interact Callbacks/AddParam")
-		.add_arg("effect_ref", effect_ref as usize)
-		.add_arg("index", index)
-		.add_arg("def", format!("{:?}", def))
-		.set_result(0)
-		.emit();
-
-	PF_Err_NONE as PF_Err
-}
-
-unsafe extern "C" fn CheckoutParam_sys(
-	effect_ref: PF_ProgPtr,
-	index: PF_ParamIndex,
-	what_time: A_long,
-	time_step: A_long,
-	time_scale: A_u_long,
-	param: *mut PF_ParamDef,
-) -> PF_Err {
-	#[cfg(feature = "diagnostics")]
-	DiagnosticBuilder::new()
-		.set_name("InteractCallbacks/RegisterUI")
-		.add_arg("effect_ref", effect_ref as usize)
-		.add_arg("index", index)
-		.add_arg("what_time", what_time)
-		.add_arg("time_step", time_step)
-		.add_arg("time_scale", time_scale)
-		.add_arg("param", format! {"{:?}", param})
-		.set_result(0)
-		.emit();
-
-	PF_Err_NONE as PF_Err
-}
-
-unsafe extern "C" fn checkin_param_sys(effect_ref: PF_ProgPtr, param: *mut PF_ParamDef) -> PF_Err {
-	#[cfg(feature = "diagnostics")]
-	DiagnosticBuilder::new()
-		.set_name("InteractCallbacks/CheckinParam")
-		.add_arg("effect_ref", effect_ref as usize)
-		.add_arg("param", format! {"{:?}", param})
-		.set_result(0)
-		.emit();
-
-	PF_Err_NONE as PF_Err
-}
-
-unsafe extern "C" fn RegisterUI_sys(
-	effect_ref: PF_ProgPtr,
-	cust_info: *mut PF_CustomUIInfo,
-) -> PF_Err {
-	#[cfg(feature = "diagnostics")]
-	DiagnosticBuilder::new()
-		.set_name("InteractCallbacks/RegisterUI")
-		.add_arg("effect_ref", effect_ref as usize)
-		.add_arg("cust_info", format!("{:?}", cust_info))
-		.set_result(0)
-		.emit();
-
-	PF_Err_NONE as PF_Err
-}
-
-unsafe extern "C" fn GetPlatformData_sys(
-	effect_ref: PF_ProgPtr,
-	which: PF_PlatDataID,
-	data: *mut ::std::os::raw::c_void,
-) -> PF_Err {
-	#[cfg(feature = "diagnostics")]
-	DiagnosticBuilder::new()
-		.set_name("UtilityCallbacks/GetPlatformData")
-		.add_arg("effect_ref", effect_ref as usize)
-		.add_arg("which", which)
-		.add_arg("data", format!("{:?}", data))
-		.set_result(0)
-		.emit();
-
-	PF_Err_NONE as PF_Err
-}
 
 /// Wrapper for After Effects plugin entry point \
 /// Note: EffectMain naming is required by the AE API and cannot be changed
@@ -103,7 +17,7 @@ unsafe extern "C" fn GetPlatformData_sys(
 #[repr(C)]
 pub struct EffectMainApi {
 	#[allow(non_snake_case)]
-	EffectMain: unsafe extern "C" fn(
+	EntryPointFunc: unsafe extern "C" fn(
 		cmd: after_effects::RawCommand,
 		in_data: *mut after_effects_sys::PF_InData,
 		out_data: *mut after_effects_sys::PF_OutData,
@@ -139,10 +53,10 @@ impl PluginInstance {
 		let width = 1920;
 		let height = 1080;
 		// Initialize Interact Callbacks using factory
-		let interact_callbacks = crate::suites::factories::create_interact_callbacks();
+		let interact_callbacks = crate::host::interact::create_interact_callbacks();
 
 		// Initialize Utility Callbacks using factory
-		let utility_callbacks = crate::suites::factories::create_utility_callbacks();
+		let utility_callbacks = crate::host::utility::create_utility_callbacks();
 
 		let input_layer = wrapper::Layer::<wrapper::Depth8>::new(
 			width,
@@ -215,90 +129,11 @@ impl PluginInstance {
 			cmd: after_effects::RawCommand::About,
 			utility_callbacks,
 			pica,
-			in_data: after_effects_sys::PF_InData {
-				inter: interact_callbacks,
-				utils: null_mut(), // Will be set later
-				effect_ref: std::ptr::null_mut(),
-				quality: after_effects_sys::PF_Quality_HI,
-				version: after_effects_sys::PF_SpecVersion {
-					major: 13,
-					minor: 28,
-				},
-				serial_num: -2147483648,
-				appl_id: 1180193859,
-				num_params: 0,
-				reserved: 0,
-				what_cpu: 3,
-				what_fpu: 0,
-				current_time: 0,
-				time_step: 1024,
-				total_time: 0,
-				local_time_step: 0,
-				time_scale: 0,
-				field: PF_Field_UPPER as PF_Field,
-				shutter_angle: 0,
-				width: 1280,
-				height: 720,
-				extent_hint: after_effects_sys::PF_UnionableRect {
-					left: 0,
-					top: 0,
-					right: 1280,
-					bottom: 720,
-				},
-				output_origin_x: 0,
-				output_origin_y: 0,
-				downsample_x: after_effects_sys::PF_RationalScale { num: 1, den: 1 }, // Fixed: den should not be 0
-				downsample_y: after_effects_sys::PF_RationalScale { num: 1, den: 1 }, // Fixed: den should not be 0
-				pixel_aspect_ratio: after_effects_sys::PF_RationalScale { num: 1, den: 1 }, // Fixed: den should not be 0
-				in_flags: PF_InFlag_NONE as PF_InFlags,
-				global_data: null_mut(),
-				sequence_data: null_mut(),
-				frame_data: null_mut(),
-				start_sampL: 0,
-				dur_sampL: 0,
-				total_sampL: 0,
-				src_snd: after_effects_sys::PF_SoundWorld {
-					fi: after_effects_sys::PF_SoundFormatInfo {
-						rateF: 1.0,
-						num_channels: 2,
-						format: 16,
-						sample_size: 1024,
-					},
-					num_samples: 1024,
-					dataP: null_mut(),
-				},
-				pica_basicP: null_mut(), // Will be set to &mut instance.pica later
-				pre_effect_source_origin_x: 0,
-				pre_effect_source_origin_y: 0,
-				shutter_phase: 0,
-			},
-			out_data: after_effects_sys::PF_OutData {
-				my_version: 0,
-				name: [0; 32],
-				global_data: null_mut(),
-				num_params: 0,
-				sequence_data: null_mut(),
-				flat_sdata_size: 0,
-				frame_data: null_mut(),
-				width: 0,
-				height: 0,
-				origin: after_effects_sys::PF_Point { h: 0, v: 0 },
-				out_flags: after_effects_sys::PF_OutFlag_NONE as after_effects_sys::PF_OutFlags,
-				return_msg: [0; 256],
-				start_sampL: 0,
-				dur_sampL: 0,
-				dest_snd: after_effects_sys::PF_SoundWorld {
-					fi: after_effects_sys::PF_SoundFormatInfo {
-						rateF: 44100.0,
-						num_channels: 2,
-						format: 16,
-						sample_size: 1024,
-					},
-					num_samples: 1024,
-					dataP: null_mut(),
-				}, // Fixed: more realistic sample rate
-				out_flags2: after_effects_sys::PF_OutFlag2_NONE as after_effects_sys::PF_OutFlags2,
-			},
+			in_data: crate::core::helpers::InDataBuilder::new()
+				.with_size(1280, 720)
+				.with_callbacks(interact_callbacks)
+				.build(),
+			out_data: crate::core::helpers::OutDataBuilder::new().build(),
 			params: param_list,
 			input_layer,
 			lllllayer: wrapper::Layer::<wrapper::Depth8>::new(
@@ -307,30 +142,9 @@ impl PluginInstance {
 				vec![wrapper::Pixel::<wrapper::Depth8>::black(); (width * height) as usize],
 			)
 			.unwrap(),
-			world: after_effects_sys::PF_LayerDef {
-				reserved0: null_mut(),
-				reserved1: null_mut(),
-				world_flags: 0 as after_effects_sys::PF_WorldFlags,
-				data: null_mut(),
-				rowbytes: (width * 4) as i32,
-				width: width as i32,
-				height: height as i32,
-				extent_hint: after_effects_sys::PF_UnionableRect {
-					left: 0,
-					top: 0,
-					right: 0,
-					bottom: 0,
-				},
-				platform_ref: null_mut(),
-				reserved_long1: 0,
-				reserved_long4: null_mut(),
-				pix_aspect_ratio: after_effects_sys::PF_RationalScale { num: 1, den: 1 }, // Fixed: den should not be 0
-				reserved_long2: null_mut(),
-				origin_x: 0,
-				origin_y: 0,
-				reserved_long3: 0,
-				dephault: 0,
-			},
+			world: crate::core::helpers::LayerDefBuilder::new()
+				.with_size(width as i32, height as i32)
+				.build(),
 		};
 
 		// Now set the utils pointer to reference our owned utility_callbacks
@@ -385,7 +199,7 @@ impl PluginInstance {
 		self.container = Some(unsafe { Container::load(&module_path) }?);
 
 		// Set plugin path for get_platform_data callback
-		crate::suites::factories::set_plugin_path(std::path::Path::new(&module_path));
+		crate::host::utility::set_plugin_path(std::path::Path::new(&module_path));
 
 		log::info!("Loaded plugin {}.", "successfully".green());
 		//* -------------------------------------------- *//
@@ -409,7 +223,7 @@ impl PluginInstance {
 			.ok_or(AexloError::ContainerNotLoaded)?;
 
 		let result = unsafe {
-			container.EffectMain(
+			container.EntryPointFunc(
 				self.cmd,
 				&mut self.in_data,
 				&mut self.out_data,
@@ -426,12 +240,14 @@ impl PluginInstance {
 		);
 
 		//* ---- Check for errors ---------------------- *//
-		match result as i32 {
+		match result as u32 {
 			PF_Err_NONE => {
 				log::info!("Plugin executed {}.", "successfully".green());
 			}
 			code => {
-				return Err(AexloError::PluginExecutionFailed { code });
+				return Err(AexloError::PluginExecutionFailed {
+					code: code.try_into().unwrap(),
+				});
 			}
 		}
 		//* -------------------------------------------- *//
