--- aexlo/src/suites/factories/mod.rs
@@ -1,11 +1,0 @@
-mod handle_suite;
-mod interact_callbacks;
-mod iterate_8_suite;
-mod utility_callbacks;
-mod world_transform_suite;
-
-pub use handle_suite::create_handle_suite_1;
-pub use interact_callbacks::create_interact_callbacks;
-pub use iterate_8_suite::create_iterate_8_suite_2;
-pub use utility_callbacks::{create_utility_callbacks, set_plugin_path};
-pub use world_transform_suite::create_world_transform_suite_1;
